Initial implementation of the Keep Network SD
This implementation of Keep Network Nodes Discovery for Prometheus is
implemented based on the example from Prometheus:
https://github.com/prometheus/prometheus/blob/734772f82824db11344ea3c39a166449d0e7e468/documentation/examples/custom-sd/adapter-usage/main.go

The list of nodes is taken from source addresses that are expected to be
set to bootstrap nodes.

The port for diagnostics endpoint is unknown for the bootstrap node so
the discovery has to scan the ports at the specific range.

The scan range will be agreed for the Keep Network and all nodes running
in the network should expose their diagnostics in this range. Preferably
the nodes should use ports in order from the lower end of the range.