Use netowork port to confirm address is reachable
A client could be exposed on an address that is not publicly reachable.
We need to validate the address is reachable before we scan ports, so we
don't waist time.

The network port is the port that has to be open and we can use it to
confirm the address is reachable.